Fix matching objects when listing in test harness
When listing objects in the test harness, there was a bug where the a for loop variable was appended to the list of objects
to compare causing only the last object (alphabetically) in the namespace to be compared when checking if any objects match
the assert.

This fixes it by indexing the list instead of taking the address of the loop variable.