Clean up integration tests that use controller-runtime envtest. (#525)
Fixes #510

envtest can race when choosing the Kubernetes API port if multiple envtests are created simultaneously.

Instead of starting envtest within each test, we start envtest in a TestMain for the module. This means we do not
need to start as many envtests, as well as reducing our likelihood of racing over a Kubernetes API port.

We also retry creating envtest. If the API server does not come up, we will try twice to try to prevent any potential races.