Fix default test timeout and make default test timeout configurable at the test suite level.
The KEP defined the default assertion timeout as 300 seconds, but we actually implemented it as 10 seconds. I think 300 seconds is much too high for the default, so I've adjusted it to 30 in both the implementation and KEP.

This also makes the timeout configurable at the test suite level so that it can be configured for an entire test suite instead of just a specific assert.