Fix race with ActivePlan between Instance and PlanExecution controllers (#722)
Based on the test I wrote in #718 it became apparent where lies the race that is one of the causes of multiple deploy plan executions.

- instance is created, create event triggers reconciliation
- planexecution is created
- OV is created and triggers reconciliation
- another planexecution is created because ActivePlan is still empty
- PlanExecutionController processes added planexecutions and create link between one of them and instance leaving two deploy planexecutions in progress, only one linked to instance, the other just hanging around forever

This PR proposes to move adding the ActivePlan relation to InstanceController. To make that possible I needed to add some safeguards to the PlanExecutionController as well.