Proposal: Support calling out to arbitrary commands in the test harness. (#754)
Currently the test harness only supports running kubectl commands and plugins, but this is limiting:

* It prevents testing Helm charts (you cannot use Helm).
* It prevents using kind subcommands (e.g., `kind load docker-image`).

This would have us support the `Commands` setting to allow this flexibility.

The downside is that now tests can call arbitrary commands. Alternatively, we could retain the kubectl restriction and require helm/kind/etc to be installed as kubectl plugins in order to be used in tests.