Introduce a designate `ResourceSpec` type
**What this PR does / why we need it**:

This is the first step towards pipe-tasks described in KEP-0017. We introduce a new `ResourceSpec` type and replace a list of task resources `Resources []string` with it `Resources []ResourceSpec`. Currently `ResourceSpec` only has one field `Name` but it is to be extended.

Note: this is a breaking change, making previous operator definitions incompatible.

Related: #774