Make Instance name deterministic
Remove the random string suffixed onto the end of instance names,
instead using a suffix of `-instance`.  Users can use the `--instance`
option to provide a unique name if they so wish.

The removal of the random string suffixed onto instance names means
that the associated test can be simplified.