Make sure config/crds are always up to date for testing (#977)
What this PR does / why we need it:
In the past, we generated manifests using /config/manifests. Then we switched to kudo init. Now kudo init should be the ultimate source of truth but at the same time, we still have /config/manifests file with usually obsolete manifests that are causing cryptic test failures.

This PR removes this tech debt. Instead of removing /config/crds, I went the way of generating them from init in tests and then writing test that always ensures that these are up to date with what init provides. The manifests are right now used e.g. when test harness starts - the harness expects folder with CRDs so this was one solution of how to provide it (I find it the easier one).