Move enhancer to renderer package (#993)
What this PR does / why we need it:
This is a follow-up of #992 where we wanted to move enhancer into the rendering package but just did not manage to do that. This PR moves it there plus the metadata from tasker, that is actually rendering metadata so it belongs there.

This also moves template package and health package into the engine and does a bit of renaming.