Remove a problematic test.
When `go test` itself is being run on a (kind) cluster, this test fails
with:
```
W1119 10:50:47.569598   12523 client_config.go:541] Neither --kubeconfig nor --master was specified.  Using the inClusterConfig.  This might not work.
--- FAIL: TestNewK2oClient (0.01s)
     kudo_test.go:32: assertion failed: expected error to contain "invalid configuration: no configuration has been provided", got operators: customresourcedefinitions.apiextensions.k8s.io "operators.kudo.dev" is forbidden: User "system:serviceaccount:dispatch:local-runner" cannot get resource "customresourcedefinitions" in API group "apiextensions.k8s.io" at the cluster scope
```

This is because of [this special case in `BuildConfigFromFlags`](https://github.com/kubernetes/client-go/blob/571c0ef67034a5e72b9e30e662044b770361641e/tools/clientcmd/client_config.go#L542-L546)
which falls back to `restclient.InClusterConfig()` when available.
Regardless of whether this config is usable in a `dispatch` kind cluster
or not, this special case proves the assumption that this test makes
("it is not possible to create a client without a config") wrong in
general.

Therefore removing this test altogether seems like the best course of
action.