Run a cleanup plan when deleting instances
If an operator provides a 'cleanup' plan, this plan is run when the operator's instance is deleted from the cluster.
This is achieved by detecting an instance deletion in the controller and adding a finalizer.
If there is another plan currently in progress, this will be superseded by the cleanup plan.
This allows the deletion of instances where a plan fails to complete.