Disallow only spec updates for now (#1146)
What this PR does / why we need it:
For now, we disallowed all updates except status updates. This narrows it down and just disallows updates to the spec that means:

- no change in params
- no change in OV
This potentially narrows down issues with the webhook as controller etc. can update annotation or finalizers, which I think is fine even when plan is running.

Addition to #883