Update plan status output to include last updated / last finished times
Introduce a new field into AggregatedStatus that contains the time which
the status was last updated.  This is then displayed via the CLI when a
user runs the `plan status` command for a particular instance.

Also updated `plan status` to show the value for the last finished run
for a given plan.

Signed-off-by: Nick Jones <nick@dischord.org>