Add integration tests for the instance admission controller (#1442)
Summary:
the tests discovered a bug in the instance admission controller (IAC) where IAC was trying to reset `PlanExecution.PlanName` when a plan becomes terminal. However, it wasn't doing this because:

1. it wasn't notified when `Instance.Status` was being updated (_Status_ is a sub-resource) and
2. even if it would, `Instance` resource **can not be changed** upon `Status` subresource updates (any resource is only allowed to modify itself from the webhook)

To fix this problem we would either need to merge `Status` back into the `Instance` resource or duplicate part of the `Status` information in the `Instance.Spec`. After long discussions, we decided to introduce a new `PlanExecution.Status` field which is populated by the instance controller and duplicates the corresponding `Status.PlanStatus` field.

Signed-off-by: Aleksey Dukhovniy <alex.dukhovniy@googlemail.com>