Handle changed dependencies when upgrading operators
When upgrading operators with dependencies, its dependencies could have changed. In this case the Operator and OperatorVersion resources of the new dependencies are installed. Outdated OperatorVersion resources aren't removed.

The code was refactored to decouple dependency handling from package installation and have separate packages to install and upgrade KUDO resources.

Signed-off-by: Jan Schlicht <jan@d2iq.com>