Local and relative dependencies are now relative to their operators
Summary:
Previously the location of a dependency operator was always relative to the caller working director (the directory of the `kudo install ...` call). This made installing and testing operators with dependencies a brittle process as the working directory has to be set accordingly. This PR fixes this for *local operators in directories:* so with an operator structure like this:

```
.
├── child
│   └── operator.yaml
│
└── parent
    └── operator.yaml
```

child operator now has to be referenced like:

```
  - name: child
    kind: KudoOperator
    spec:
      package: "../child"
```

*Note:* This fix does not apply to tarballs as the tarball reader does not expect any dependency operators files/folders. As for right now KUDO does not support packaging/extracting operators with dependencies.

Signed-off-by: Aleksey Dukhovniy <alex.dukhovniy@googlemail.com>