Test adding a required no-default parameter on upgrade.
Currently, an operator upgrade that brings in a new required parameter
without a default, is impossible if the operator defines both
"update" and "upgrade" plans.

An upgrade without setting this parameter fails (which is a correct
behaviour); however, an attempt to upgrade while setting the previously
nonexistent parameter also fails due to both plans being triggered,
regardless of the fact that no existing parameter is changed.

This test verifies that an upgrade of such an operator is possible
together with setting a value for the new required parameter,
and that the plan chosen is "upgrade" and not "update".

Signed-off-by: Andrei Sekretenko <asekretenko@d2iq.com>